Move AsyncFs extension functions to separate file
