+++ library/async/src/jvmMain/java9/module-info.java
@@ -1,5 +1,6 @@
 module io.matthewnelson.kmp.file.async {
     requires transitive io.matthewnelson.kmp.file;
+    requires kotlinx.coroutines.core;
 
     exports io.matthewnelson.kmp.file.async;
 }
